DL/Udemy/Generative-AI/Krish Naik/SUBHASISH BISWAS/Langchain Projects/requirements.txt
langchain
python-dotenv
ipykernel
langchain-community
pypdf
bs4
arxiv
pymupdf
wikipedia
langchain-text-splitters
langchain-openai
chromadb
sentence_transformers
langchain_huggingface
faiss-cpu
langchain_chroma
duckdb
pandas
openai
arxiv
wikipedia
beautifulsoup4
streamlit
langchain_groq
fastapi
uvicorn
langserve
sse_starlette